+++ src/SkeletonClass.php
@@ -1,5 +1,0 @@
-<?php
-
-namespace VendorName\Skeleton;
-
-class SkeletonClass {}